requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
groq>=0.5.0
google-genai>=0.3.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
